+++ packages/res/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/res
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/send@0.0.3
+
 ## 0.0.3
 
 ### Patch Changes